Make proxy tensor support item() calls on torch.tensor constants
Read the comment in torch/fx/experimental/proxy_tensor.py

This is mildly BC-breaking if anyone was previously interposing on
at::lift, but this operator was relatively new and I checked
functorch which has no explicit reference to lift.  So I think it
should not be too disruptive.

There are some extra modifications to fake tensor: the change
that is substantive to this change is changing the lift special
case to handle lift_fresh and lift_fresh_copy, but while I'm at
it I fix a bug where I make fake tensor mode defer to other
subclasses if there are other subclasses in the arguments.

Signed-off-by: Edward Z. Yang <ezyangfb.com>

ghstack-source-id: b3eeee90954554e246b4aa3f78e92cfa739e7353
Pull Request resolved: https://github.com/pytorch/pytorch/pull/81192